pyproject.toml
[tool.poetry]
name = "interedit"
version = "0.1.0" # PACKAGE_VERSION  (This comment helps bumpversion find the right line.)
description = ""
authors = ["Interdoc contributors <interdoc-edit-bot@users.noreply.github.com>"]
readme = "README.rst"

[tool.poetry.dependencies]
python = "^3.7"
attrs = "19.1.0"
pyramid = "^1.10"
shiv = "^0.0.48"
click = "<7.0"
hyperlink = "^19.0"
marshmallow = "^3.2"
fuzzywuzzy = "^0.17.0"


[tool.poetry.dev-dependencies]
bump2version = "^0.5.10"
coverage = "^4.5"
pytest = "^5.0"
hypothesis = "^4.32"
requests = "^2.22"
pylint = "^2.3"
aiohttp = "^3.5"
twine = "^1.13"
vulture = "^1.0"
cuvner = "^18.0"
pytest-cov = "^2.7"
tox = "^3.13"
sphinx_rtd_theme = "^0.4.3"
pytest-sugar = "^0.9.2"
docshtest = "^0.0.2"
pre-commit = "^1.18"
black = {version = "^18.3-alpha.0", allows-prereleases = true}
pytest-clarity = {version = "^0.1.0-alpha.1", allows-prereleases = true}
pygments = {version = "^2.4", allows-prereleases = true}
pudb = "^2019.1"

[tool.poetry.extras]
docs = [
    "sphinx",
    "sphinx-rtd-theme",
    "sphinx-autodoc-typehints",
    "sphinx-click",
    "marshmallow-jsonschema",
    "sphinx-jsonschema",

]


[tool.poetry.scripts]
interedit = "interedit.cli:cli"


[tool.towncrier]

   package = "interedit"
   package_dir = "src"
   filename = "CHANGELOG.rst"
   template = "changelog.d/towncrier_template.rst"
   issue_format = "`#{issue} <https://github.com/interdoc-edit-bot/interedit/issues/{issue}>`_"
   directory = "changelog.d"
   title_format = "{version} ({project_date})"
   underlines = ["-", "^"]

   [[tool.towncrier.section]]
       path = ""

    [[tool.towncrier.type]]
        directory = "breaking"
        name = "Backward-incompatible Changes"
        showcontent = true

    [[tool.towncrier.type]]
        directory = "deprecation"
        name = "Deprecations"
        showcontent = true

    [[tool.towncrier.type]]
        directory = "change"
        name = "Changes"
        showcontent = true


[tool.black]

include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.?venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
